Refactoring to follow template
