+++ packages/mds-compliance-service/repository/mappers.ts
@@ -1,7 +1,7 @@
 import { Optional } from '@mds-core/mds-types'
 import { RecordedColumn, IdentityColumn, ModelMapper } from '@mds-core/mds-repository'
-import { ComplianceResponsePersistenceModel } from './entities/compliance-response-entity'
-import { ComplianceResponseDomainModel, ComplianceResponseDomainCreateModel } from '../../../@types'
+import { ComplianceResponsePersistenceModel } from './entities'
+import { ComplianceResponseDomainModel } from '../@types'
 
 export const ComplianceResponsePersistenceToDomain = ModelMapper<
   ComplianceResponsePersistenceModel,
@@ -11,14 +11,10 @@ export const ComplianceResponsePersistenceToDomain = ModelMapper<
   return domain
 })
 
+/*
 type ComplianceResponsePersistenceCreateModel = Omit<
   Optional<ComplianceResponsePersistenceModel, keyof RecordedColumn>,
   keyof IdentityColumn
 >
 
-export const ComplianceResponseDomainToEntityCreate = ModelMapper<
-  ComplianceResponseDomainCreateModel,
-  ComplianceResponsePersistenceCreateModel
->((domain, options) => {
-  return domain
-})
+*/
